--- app/src/main/java/com/example/thefalgbusstop/domain/entities/EntityBeingModified.kt
@@ -1,17 +1,0 @@
-package com.example.thefalgbusstop.domain.entities
-
-import androidx.databinding.BaseObservable
-import androidx.databinding.Bindable
-
-data class EntityBeingModified(
-    var editTextString1: String? = null,
-    var editTextString2: String? = null,
-    var editTextString3: String? = null,
-    var editTextInt1: String? = null,
-    var editTextInt2: String? = null,
-    var editTextInt3: String? = null,
-
-
-    ) {
-
-}
