+++ app/src/main/java/com/example/thefalgbusstop/presentation/activities/MainActivity.kt
@@ -1,28 +1,33 @@
 package com.example.thefalgbusstop.presentation.activities
 
+import android.content.Intent
 import android.os.Bundle
 import android.util.Log
+import android.view.Menu
 import android.view.MenuItem
-import androidx.appcompat.widget.Toolbar
+import android.widget.Toast
 import androidx.appcompat.app.AppCompatActivity
+import androidx.appcompat.widget.Toolbar
 import androidx.databinding.DataBindingUtil
 import androidx.navigation.fragment.NavHostFragment
 import androidx.navigation.ui.NavigationUI
 import com.example.thefalgbusstop.R
-import com.example.thefalgbusstop.utils.Cosntants
+import com.example.thefalgbusstop.databinding.MainActivityBinding
 import com.example.thefalgbusstop.domain.entities.Bus
 import com.example.thefalgbusstop.domain.entities.Chofer
 import com.example.thefalgbusstop.domain.entities.Horarios
 import com.example.thefalgbusstop.domain.entities.Passenger
-import com.example.thefalgbusstop.utils.startActivity
-import com.example.thefalgbusstop.databinding.MainActivityBinding
 import com.example.thefalgbusstop.parcelables.toChoferParcelable
-import com.example.thefalgbusstop.presentation.Fragments.Buses.BusesFragment
-import com.example.thefalgbusstop.presentation.Fragments.Chofers.List.ChoferListFragment
-import com.example.thefalgbusstop.presentation.Fragments.Hours.HoursFragment
-import com.example.thefalgbusstop.presentation.Fragments.Passengers.PassengersFragment
+import com.example.thefalgbusstop.presentation.fragments.Buses.BusesFragment
+import com.example.thefalgbusstop.presentation.fragments.Chofers.List.ChoferListFragment
+import com.example.thefalgbusstop.presentation.fragments.Hours.HoursFragment
+import com.example.thefalgbusstop.presentation.fragments.Passengers.PassengersFragment
+import com.example.thefalgbusstop.utils.Cosntants
+import com.example.thefalgbusstop.utils.Utils
+import com.example.thefalgbusstop.utils.startActivity
 import com.google.android.material.bottomnavigation.BottomNavigationView
 import com.google.android.material.navigation.NavigationView
+import kotlinx.android.synthetic.main.main_activity.*
 
 
 class MainActivity : AppCompatActivity(),
@@ -32,7 +37,6 @@ class MainActivity : AppCompatActivity(),
     PassengersFragment.OnPassengersFragmentListener,
     NavigationView.OnNavigationItemSelectedListener{
     lateinit var bottomNavigationView: BottomNavigationView
-    lateinit var toolbar: Toolbar
     private lateinit var  binding: MainActivityBinding
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -42,35 +46,65 @@ class MainActivity : AppCompatActivity(),
     }
 
     override fun onNavigationItemSelected(item: MenuItem): Boolean {
-        when(item.itemId){
-            R.id.chofersFragment,R.id.busesFragment, R.id.hoursFragment -> {
+        return when(item.itemId){
+            R.id.chofersFragment, R.id.busesFragment, R.id.hoursFragment -> {
                 binding.title = item.title.toString()
-                return true
+                true
             }
             else -> {
                 binding.title = "Estacion la Bandera"
-                return false
+                false
             }
 
         }
     }
 
+    override fun onCreateOptionsMenu(menu: Menu?): Boolean {
+        Log.i("MAin", "onCreateOptionsMenu: estamos en el menu")
+        menuInflater.inflate(R.menu.toolbar_menu, menu)
+        return true
+    }
 
-    fun setUpNavigation(){
+    //Handling Action Bar button click
+    override fun onOptionsItemSelected(item: MenuItem): Boolean {
+        // Handle item selection
+        Log.i("MAin", "onOptionsItemSelected: ${item.itemId}")
+        val id = item.itemId
+        Log.i("MAin", "onOptionsItemSelected: ${item.itemId}")
+        return when (id) {
+            R.id.add_entity -> {
+                Toast.makeText(this, "clicl en agregar", Toast.LENGTH_SHORT).show()
+                addEntity()
+                true
+            }
+            else -> super.onOptionsItemSelected(item)
+        }
+    }
+
+
+    private fun setUpNavigation(){
         bottomNavigationView = findViewById(R.id.bottomNavigationView)
         val navHostFragment = supportFragmentManager
             .findFragmentById(R.id.nav_host_fragment) as NavHostFragment
         NavigationUI.setupWithNavController(
-                bottomNavigationView,
-                navHostFragment.navController
+            bottomNavigationView,
+            navHostFragment.navController
 
         )
     }
 
     private fun initComponents(){
-        toolbar = findViewById(R.id.toolbar)
+        title = "Terminal la Bandera App"
         setUpNavigation()
     }
+
+    private fun addEntity(){
+        val intent = Intent(this, EditItemActivity::class.java)
+        startActivity(intent)
+        overridePendingTransition(R.anim.entry, R.anim.exit)
+    }
+
+
 
     override fun openChoferDetail(chofer: Chofer) {
         startActivity<ItemDetailActivity> {
@@ -80,15 +114,28 @@ class MainActivity : AppCompatActivity(),
     }
 
     override fun openBusDetail(Bus: Bus) {
-        Log.i("MainActv", "openChoferDetail: Aun no hay pantalla de detalle")
+        notYetImplementDialog()
     }
 
     override fun openHoursDetail(Hours: Horarios) {
-        Log.i("MainActv", "openChoferDetail: Aun no hay pantalla de detalle")
+        notYetImplementDialog()
     }
 
     override fun openPassengerDetail(passenger: Passenger) {
-        Log.i("MainActv", "openChoferDetail: Aun no hay pantalla de detalle")
+        notYetImplementDialog()
+    }
+
+    private fun notYetImplementDialog(){
+        val nDialog = Utils.confirmDialof(
+            this,
+            "En Desarrollo",
+            "Estamos trabajando para desarrollar todos estos features, proximamente se notificará",
+            "Continuar"
+        )
+        nDialog.setConfirmClickListener {
+            nDialog.dismissWithAnimation()
+        }
+        nDialog.show()
     }
