+++ app/src/main/res/drawable/ic_add_user.xml
@@ -4,6 +4,6 @@
     android:viewportWidth="512"
     android:viewportHeight="512">
   <path
-      android:fillColor="#FF3700B3"
+      android:fillColor="#FFFFFFFF"
       android:pathData="m49.5,512c-11.046,0 -20,-8.954 -20,-20 0,-100.355 81.645,-182 182,-182h30c20.537,0 40.703,3.4 59.937,10.105 10.43,3.636 15.938,15.04 12.301,25.469 -3.636,10.43 -15.04,15.938 -25.469,12.301 -14.991,-5.225 -30.726,-7.875 -46.769,-7.875h-30c-78.299,0 -142,63.701 -142,142 0,11.046 -8.954,20 -20,20zM359.5,135c0,-74.439 -60.561,-135 -135,-135s-135,60.561 -135,135 60.561,135 135,135 135,-60.561 135,-135zM319.5,135c0,52.383 -42.617,95 -95,95s-95,-42.617 -95,-95 42.617,-95 95,-95 95,42.617 95,95zM462.5,392h-60v-60c0,-11.046 -8.954,-20 -20,-20s-20,8.954 -20,20v60h-60c-11.046,0 -20,8.954 -20,20s8.954,20 20,20h60v60c0,11.046 8.954,20 20,20s20,-8.954 20,-20v-60h60c11.046,0 20,-8.954 20,-20s-8.954,-20 -20,-20z"/>
 </vector>
