+++ app/src/main/res/drawable/ic_autobus.xml
@@ -1,0 +1,27 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="512dp"
+    android:height="512dp"
+    android:viewportWidth="512"
+    android:viewportHeight="512">
+  <path
+      android:fillColor="#FF9e9e9e"
+      android:pathData="M157.152,315.119c-24.262,0 -44,19.738 -44,44c0,24.262 19.738,44 44,44c24.262,0 44,-19.738 44,-44C201.152,334.857 181.414,315.119 157.152,315.119zM157.152,383.119c-13.233,0 -24,-10.767 -24,-24c0,-13.233 10.767,-24 24,-24c13.233,0 24,10.767 24,24C181.152,372.352 170.385,383.119 157.152,383.119z"/>
+  <path
+      android:fillColor="#FF9e9e9e"
+      android:pathData="M457.251,108.884H406c-5.523,0 -10,4.477 -10,10c0,5.523 4.477,10 10,10h51.251c15.991,0 29.484,10.86 33.515,25.59H406c-5.523,0 -10,4.477 -10,10v89.915c0,5.523 4.477,10 10,10h86v84.914h-29.342c-4.474,-19.549 -21.996,-34.183 -42.882,-34.183s-38.408,14.633 -42.882,34.183H376V164.562c0,-5.523 -4.477,-10 -10,-10h-80c-5.523,0 -10,4.477 -10,10v80.322H90.333c-5.523,0 -10,4.477 -10,10c0,5.523 4.477,10 10,10H276v84.417h-51c-5.523,0 -10,4.477 -10,10s4.477,10 10,10h151.978c4.608,19.366 22.041,33.817 42.798,33.817s38.19,-14.451 42.798,-33.817H502c5.523,0 10,-4.477 10,-10V163.633C512,133.444 487.44,108.884 457.251,108.884zM356,251.973h-6.321c-5.523,0 -10,4.477 -10,10s4.477,10 10,10H356v77.328h-60V174.562h60V251.973zM419.775,383.119c-13.233,0 -24,-10.767 -24,-24c0,-13.233 10.767,-24 24,-24c13.233,0 24,10.767 24,24C443.775,372.352 433.009,383.119 419.775,383.119zM492,244.388h-76v-69.915h76V244.388z"/>
+  <path
+      android:fillColor="#FF9e9e9e"
+      android:pathData="M373.07,111.811c-1.86,-1.86 -4.44,-2.93 -7.07,-2.93s-5.21,1.07 -7.07,2.93c-1.86,1.86 -2.93,4.44 -2.93,7.07s1.07,5.21 2.93,7.07c1.86,1.86 4.44,2.93 7.07,2.93s5.21,-1.07 7.07,-2.93s2.93,-4.44 2.93,-7.07S374.93,113.671 373.07,111.811z"/>
+  <path
+      android:fillColor="#FF9e9e9e"
+      android:pathData="M326,108.884H54.749c-27.414,0 -50.183,20.255 -54.137,46.585C0.223,156.536 0,157.683 0,158.884v200.417c0,5.523 4.477,10 10,10h75.278c5.523,0 10,-4.477 10,-10s-4.477,-9.999 -10,-9.999H20V224.884h34c5.523,0 10,-4.477 10,-10v-56c0,-5.523 -4.477,-10 -10,-10H23.294c5.558,-11.806 17.564,-20 31.454,-20H326c5.523,0 10,-4.477 10,-10C336,113.361 331.523,108.884 326,108.884zM44,168.884v36H20v-36H44z"/>
+  <path
+      android:fillColor="#FF9e9e9e"
+      android:pathData="M61.07,247.811c-1.86,-1.86 -4.44,-2.93 -7.07,-2.93s-5.21,1.07 -7.07,2.93s-2.93,4.44 -2.93,7.07c0,2.64 1.07,5.21 2.93,7.07c1.86,1.87 4.44,2.93 7.07,2.93s5.21,-1.06 7.07,-2.93c1.86,-1.86 2.93,-4.44 2.93,-7.07S62.93,249.671 61.07,247.811z"/>
+  <path
+      android:fillColor="#FF9e9e9e"
+      android:pathData="M246,148.884h-56c-5.523,0 -10,4.477 -10,10v56c0,5.523 4.477,10 10,10h56c5.523,0 10,-4.477 10,-10v-56C256,153.361 251.523,148.884 246,148.884zM236,204.884h-36v-36h36V204.884z"/>
+  <path
+      android:fillColor="#FF9e9e9e"
+      android:pathData="M150,148.884H94c-5.523,0 -10,4.477 -10,10v56c0,5.523 4.477,10 10,10h56c5.523,0 10,-4.477 10,-10v-56C160,153.361 155.523,148.884 150,148.884zM140,204.884h-36v-36h36V204.884z"/>
+</vector>
