+++ app/src/main/res/drawable/ic_avatar.xml
@@ -1,0 +1,12 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="482.9dp"
+    android:height="482.9dp"
+    android:viewportWidth="482.9"
+    android:viewportHeight="482.9">
+  <path
+      android:fillColor="#FF9e9e9e"
+      android:pathData="M239.7,260.2c0.5,0 1,0 1.6,0c0.2,0 0.4,0 0.6,0c0.3,0 0.7,0 1,0c29.3,-0.5 53,-10.8 70.5,-30.5c38.5,-43.4 32.1,-117.8 31.4,-124.9c-2.5,-53.3 -27.7,-78.8 -48.5,-90.7C280.8,5.2 262.7,0.4 242.5,0h-0.7c-0.1,0 -0.3,0 -0.4,0h-0.6c-11.1,0 -32.9,1.8 -53.8,13.7c-21,11.9 -46.6,37.4 -49.1,91.1c-0.7,7.1 -7.1,81.5 31.4,124.9C186.7,249.4 210.4,259.7 239.7,260.2zM164.6,107.3c0,-0.3 0.1,-0.6 0.1,-0.8c3.3,-71.7 54.2,-79.4 76,-79.4h0.4c0.2,0 0.5,0 0.8,0c27,0.6 72.9,11.6 76,79.4c0,0.3 0,0.6 0.1,0.8c0.1,0.7 7.1,68.7 -24.7,104.5c-12.6,14.2 -29.4,21.2 -51.5,21.4c-0.2,0 -0.3,0 -0.5,0l0,0c-0.2,0 -0.3,0 -0.5,0c-22,-0.2 -38.9,-7.2 -51.4,-21.4C157.7,176.2 164.5,107.9 164.6,107.3z"/>
+  <path
+      android:fillColor="#FF9e9e9e"
+      android:pathData="M446.8,383.6c0,-0.1 0,-0.2 0,-0.3c0,-0.8 -0.1,-1.6 -0.1,-2.5c-0.6,-19.8 -1.9,-66.1 -45.3,-80.9c-0.3,-0.1 -0.7,-0.2 -1,-0.3c-45.1,-11.5 -82.6,-37.5 -83,-37.8c-6.1,-4.3 -14.5,-2.8 -18.8,3.3c-4.3,6.1 -2.8,14.5 3.3,18.8c1.7,1.2 41.5,28.9 91.3,41.7c23.3,8.3 25.9,33.2 26.6,56c0,0.9 0,1.7 0.1,2.5c0.1,9 -0.5,22.9 -2.1,30.9c-16.2,9.2 -79.7,41 -176.3,41c-96.2,0 -160.1,-31.9 -176.4,-41.1c-1.6,-8 -2.3,-21.9 -2.1,-30.9c0,-0.8 0.1,-1.6 0.1,-2.5c0.7,-22.8 3.3,-47.7 26.6,-56c49.8,-12.8 89.6,-40.6 91.3,-41.7c6.1,-4.3 7.6,-12.7 3.3,-18.8c-4.3,-6.1 -12.7,-7.6 -18.8,-3.3c-0.4,0.3 -37.7,26.3 -83,37.8c-0.4,0.1 -0.7,0.2 -1,0.3c-43.4,14.9 -44.7,61.2 -45.3,80.9c0,0.9 0,1.7 -0.1,2.5c0,0.1 0,0.2 0,0.3c-0.1,5.2 -0.2,31.9 5.1,45.3c1,2.6 2.8,4.8 5.2,6.3c3,2 74.9,47.8 195.2,47.8s192.2,-45.9 195.2,-47.8c2.3,-1.5 4.2,-3.7 5.2,-6.3C447,415.5 446.9,388.8 446.8,383.6z"/>
+</vector>
