+++ app/src/main/res/drawable/ic_bus.xml
@@ -1,0 +1,9 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="405.333dp"
+    android:height="405.333dp"
+    android:viewportWidth="405.333"
+    android:viewportHeight="405.333">
+  <path
+      android:fillColor="#FF9e9e9e"
+      android:pathData="M202.667,0C108.373,0 32,10.667 32,85.333v213.333c0,18.88 8.32,35.627 21.333,47.36V384c0,11.733 9.6,21.333 21.333,21.333H96c11.84,0 21.333,-9.6 21.333,-21.333v-21.333H288V384c0,11.733 9.493,21.333 21.333,21.333h21.333c11.733,0 21.333,-9.6 21.333,-21.333v-37.973c13.013,-11.733 21.333,-28.48 21.333,-47.36V85.333C373.333,10.667 296.96,0 202.667,0zM106.667,320c-17.707,0 -32,-14.293 -32,-32s14.293,-32 32,-32s32,14.293 32,32S124.373,320 106.667,320zM298.667,320c-17.707,0 -32,-14.293 -32,-32s14.293,-32 32,-32s32,14.293 32,32S316.373,320 298.667,320zM330.667,192h-256V85.333h256V192z"/>
+</vector>
