+++ app/src/main/res/drawable/ic_bus_schedule.xml
@@ -1,0 +1,24 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="512dp"
+    android:height="512dp"
+    android:viewportWidth="60"
+    android:viewportHeight="60">
+  <path
+      android:fillColor="#FF9e9e9e"
+      android:pathData="m57.86,8c-2.1211,-3.6765 -5.6162,-6.3596 -9.716,-7.4586 -4.0998,-1.099 -8.4683,-0.524 -12.144,1.5986l-0.08,0.07c-0.5258,0.3033 -1.0334,0.6372 -1.52,1 -4.5538,-1.9711 -9.4402,-3.0603 -14.4,-3.21 -0.4,0 -9.94,0.06 -18.11,5.19 -1.1788,0.7319 -1.8941,2.0225 -1.89,3.41v48.4c0,1.6569 1.3431,3 3,3h4c1.6569,0 3,-1.3431 3,-3v-6h20v6c0,1.6569 1.3431,3 3,3h4c1.6569,0 3,-1.3431 3,-3v-25.53c13.73,3.53 24.92,-11.22 17.86,-23.47zM8,57c0,0.5523 -0.4477,1 -1,1h-4c-0.5523,0 -1,-0.4477 -1,-1v-6h6zM38,57c0,0.5523 -0.4477,1 -1,1h-4c-0.5523,0 -1,-0.4477 -1,-1v-6h6zM38,49h-36v-40.4c0.0078,-0.7066 0.388,-1.3567 1,-1.71 7.62,-4.82 16.91,-4.89 17,-4.89 4.3594,0.1545 8.6594,1.0611 12.71,2.68 -0.9678,0.9693 -1.8079,2.0581 -2.5,3.24l-0.07,0.08v0.08c-0.5401,0.9263 -0.9856,1.9046 -1.33,2.92h-21.81c-1.6223,-0.0007 -2.9513,1.2884 -3,2.91v18.18c0.0487,1.6216 1.3777,2.9107 3,2.91h26c1.6223,0.0007 2.9513,-1.2884 3,-2.91v-2.23h0.11c0.6077,0.3505 1.2393,0.6579 1.89,0.92zM30.14,24c0.0309,0.0452 0.0643,0.0886 0.1,0.13 0.9865,1.6579 2.2596,3.1274 3.76,4.34v3.62c-0.0467,0.5169 -0.481,0.9121 -1,0.91h-26c-0.519,0.0021 -0.9533,-0.3931 -1,-0.91v-18.18c0.0467,-0.5169 0.481,-0.9121 1,-0.91h21.29c-0.7228,3.7699 -0.0662,7.674 1.85,11zM51.85,27.59 L51.37,26.76c-0.2761,-0.4805 -0.8895,-0.6461 -1.37,-0.37s-0.6461,0.8895 -0.37,1.37l0.48,0.82c-1.5973,0.7946 -3.3317,1.2766 -5.11,1.42v-1c0,-0.5523 -0.4477,-1 -1,-1s-1,0.4477 -1,1v1c-1.7749,-0.1278 -3.5092,-0.5927 -5.11,-1.37l0.48,-0.82c0.2761,-0.4805 0.1105,-1.0939 -0.37,-1.37s-1.0939,-0.1105 -1.37,0.37l-0.48,0.83c-1.4707,-1.0005 -2.7395,-2.2693 -3.74,-3.74l0.83,-0.48c0.4805,-0.2761 0.6461,-0.8895 0.37,-1.37s-0.8895,-0.6461 -1.37,-0.37l-0.82,0.48c-0.7834,-1.6162 -1.2486,-3.368 -1.37,-5.16h0.95c0.5523,0 1,-0.4477 1,-1s-0.4477,-1 -1,-1h-0.95c0.1278,-1.7749 0.5927,-3.5092 1.37,-5.11l0.82,0.48c0.4805,0.2761 1.0939,0.1105 1.37,-0.37s0.1105,-1.0939 -0.37,-1.37l-0.83,-0.48c1.0005,-1.4707 2.2693,-2.7395 3.74,-3.74l0.48,0.83c0.2761,0.4805 0.8895,0.6461 1.37,0.37s0.6461,-0.8895 0.37,-1.37l-0.48,-0.82c3.8528,-1.8956 8.3672,-1.8956 12.22,0l-0.48,0.82c-0.1786,0.3108 -0.1779,0.6933 0.002,1.0034s0.5115,0.5007 0.87,0.5 0.6894,-0.1926 0.868,-0.5034l0.48,-0.83c1.4707,1.0005 2.7395,2.2693 3.74,3.74l-0.83,0.48c-0.4805,0.2761 -0.6461,0.8895 -0.37,1.37s0.8895,0.6461 1.37,0.37l0.82,-0.48c0.7946,1.5973 1.2766,3.3317 1.42,5.11h-1c-0.5523,0 -1,0.4477 -1,1s0.4477,1 1,1h1c-0.1278,1.7749 -0.5927,3.5092 -1.37,5.11l-0.82,-0.48c-0.4805,-0.2761 -1.0939,-0.1105 -1.37,0.37s-0.1105,1.0939 0.37,1.37l0.83,0.48c-1.0153,1.4753 -2.3014,2.7444 -3.79,3.74z"/>
+  <path
+      android:fillColor="#FF9e9e9e"
+      android:pathData="m46.7,17.29c0.363,-0.7656 0.3854,-1.649 0.0615,-2.432 -0.3239,-0.783 -0.9637,-1.3925 -1.7615,-1.678v-7.18c0,-0.5523 -0.4477,-1 -1,-1s-1,0.4477 -1,1v7.18c-1.0837,0.3401 -1.8824,1.2629 -2.0634,2.3842s0.2867,2.2486 1.2082,2.9124 2.139,0.7503 3.1452,0.2234l4,4c0.2537,0.2537 0.6234,0.3527 0.9699,0.2599 0.3465,-0.0928 0.6172,-0.3635 0.71,-0.71 0.0928,-0.3465 -0.0062,-0.7162 -0.2599,-0.9699zM43,16c-0.0019,-0.4526 0.3004,-0.8501 0.7372,-0.969s0.8989,0.0702 1.1268,0.4613c0.2279,0.391 0.1648,0.8864 -0.1539,1.2078 -0.2848,0.2872 -0.7146,0.3745 -1.0889,0.2213s-0.6195,-0.5168 -0.6211,-0.9213z"/>
+  <path
+      android:fillColor="#FF9e9e9e"
+      android:pathData="m11,38h-4c-1.6569,0 -3,1.3431 -3,3v2c0,1.6569 1.3431,3 3,3h4c1.6569,0 3,-1.3431 3,-3v-2c0,-1.6569 -1.3431,-3 -3,-3zM12,43c0,0.5523 -0.4477,1 -1,1h-4c-0.5523,0 -1,-0.4477 -1,-1v-2c0,-0.5523 0.4477,-1 1,-1h4c0.5523,0 1,0.4477 1,1z"/>
+  <path
+      android:fillColor="#FF9e9e9e"
+      android:pathData="m33,38h-4c-1.6569,0 -3,1.3431 -3,3v2c0,1.6569 1.3431,3 3,3h4c1.6569,0 3,-1.3431 3,-3v-2c0,-1.6569 -1.3431,-3 -3,-3zM34,43c0,0.5523 -0.4477,1 -1,1h-4c-0.5523,0 -1,-0.4477 -1,-1v-2c0,-0.5523 0.4477,-1 1,-1h4c0.5523,0 1,0.4477 1,1z"/>
+  <path
+      android:fillColor="#FF9e9e9e"
+      android:pathData="m23,42h-6c-0.5523,0 -1,0.4477 -1,1s0.4477,1 1,1h6c0.5523,0 1,-0.4477 1,-1s-0.4477,-1 -1,-1z"/>
+  <path
+      android:fillColor="#FF9e9e9e"
+      android:pathData="m15,10h10c1.1046,0 2,-0.8954 2,-2v-2c0,-1.1046 -0.8954,-2 -2,-2h-10c-1.1046,0 -2,0.8954 -2,2v2c0,1.1046 0.8954,2 2,2zM15,6h10v2h-10z"/>
+</vector>
