+++ app/src/main/res/drawable/ic_conductor.xml
@@ -1,0 +1,12 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="512dp"
+    android:height="512dp"
+    android:viewportWidth="512"
+    android:viewportHeight="512">
+  <path
+      android:fillColor="#FF9e9e9e"
+      android:pathData="m321.296,219.386c32.575,-21.352 54.137,-58.179 54.137,-99.953 0,-65.855 -53.578,-119.433 -119.433,-119.433s-119.433,53.578 -119.433,119.433c0,41.774 21.563,78.601 54.137,99.953 -82.69,27.439 -142.504,105.501 -142.504,197.281v32.133c0,34.849 28.351,63.2 63.2,63.2h289.2c34.849,0 63.2,-28.352 63.2,-63.2v-32.133c0,-91.78 -59.814,-169.842 -142.504,-197.281zM166.567,119.433c0,-49.313 40.12,-89.433 89.433,-89.433s89.433,40.12 89.433,89.433c0,49.314 -40.12,89.434 -89.433,89.434s-89.433,-40.12 -89.433,-89.434zM256,238.867c77.453,0 143.485,49.787 167.806,119.033 -7.19,-2.849 -15.014,-4.433 -23.206,-4.433 -5.039,0 -9.951,0.798 -14.608,2.309 -22.77,-48.701 -72.416,-82.643 -129.992,-82.643 -57.464,0 -107.174,33.841 -129.992,82.643 -4.657,-1.512 -9.57,-2.309 -14.608,-2.309 -8.192,0 -16.016,1.584 -23.206,4.433 24.321,-69.247 90.353,-119.033 167.806,-119.033zM128.533,464.867c0,9.447 -7.686,17.133 -17.133,17.133 -18.306,0 -33.2,-14.894 -33.2,-33.2v-32.133c0,-18.306 14.894,-33.2 33.2,-33.2 9.488,0 17.133,7.676 17.133,17.133zM241,482h-85.704c2.082,-5.315 3.238,-11.09 3.238,-17.133v-33.2h52.795c4.694,13.94 15.732,24.978 29.672,29.672v20.661zM256,433.8c-9.447,0 -17.133,-7.686 -17.133,-17.133s7.686,-17.133 17.133,-17.133 17.133,7.686 17.133,17.133 -7.686,17.133 -17.133,17.133zM271,482v-20.662c13.94,-4.694 24.978,-15.732 29.672,-29.672h52.795v33.2c0,6.043 1.156,11.819 3.238,17.133h-85.705zM353.467,400.6v1.066h-52.795c-6.281,-18.655 -23.926,-32.133 -44.672,-32.133s-38.391,13.478 -44.672,32.133h-52.795v-1.066c0,-9.457 -2.767,-18.489 -7.918,-26.172 14.422,-35.751 51.557,-71.295 105.385,-71.295 53.76,0 90.934,35.471 105.385,71.295 -5.151,7.683 -7.918,16.715 -7.918,26.172zM433.8,448.8c0,18.307 -14.894,33.2 -33.2,33.2 -9.447,0 -17.133,-7.686 -17.133,-17.133v-64.267c0,-9.449 7.638,-17.133 17.133,-17.133 18.306,0 33.2,14.894 33.2,33.2z"/>
+  <path
+      android:fillColor="#FF9e9e9e"
+      android:pathData="m256,158.505c14.128,0 28.257,-5.378 39.012,-16.134 5.858,-5.858 5.858,-15.355 0,-21.213 -5.857,-5.858 -15.355,-5.858 -21.213,0 -9.814,9.814 -25.784,9.814 -35.599,0 -5.857,-5.858 -15.355,-5.858 -21.213,0s-5.858,15.355 0,21.213c10.756,10.756 24.885,16.134 39.013,16.134z"/>
+</vector>
