+++ app/src/main/res/layout/chofers_detail_fragment.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".presentation.Fragments.Chofers.Detail.ChofersDetailFragment">
+    tools:context=".presentation.fragments.Chofers.Detail.ChofersDetailFragment">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
