+++ app/src/main/res/layout/hours_fragment.xml
@@ -11,7 +11,7 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:orientation="vertical"
-        tools:context=".presentation.Fragments.Hours.HoursFragment">
+        tools:context=".presentation.fragments.Hours.HoursFragment">
 
         <androidx.swiperefreshlayout.widget.SwipeRefreshLayout
             android:id="@+id/srwHours"
