+++ app/src/main/res/layout/route_list_fragment.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".presentation.Fragments.Routes.RouteList">
+    tools:context=".presentation.fragments.Routes.RouteList">
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent">
