+++ app/src/main/res/menu/bottom_navigation.xml
@@ -5,24 +5,24 @@
         android:id="@+id/chofersFragment"
         app:showAsAction="always"
         android:orderInCategory="2"
-        android:icon="@drawable/ic_launcher_foreground"
+        android:icon="@drawable/ic_conductor"
         android:title="Choferes" />
     <item
         android:id="@+id/hoursFragment"
         app:showAsAction="always"
         android:orderInCategory="2"
-        android:icon="@android:drawable/ic_menu_month"
+        android:icon="@drawable/ic_bus_schedule"
         android:title="Horarios" />
     <item
         android:id="@+id/busesFragment"
         app:showAsAction="always"
         android:orderInCategory="2"
-        android:icon="@android:drawable/ic_lock_idle_charging"
+        android:icon="@drawable/ic_autobus"
         android:title="Buses" />
     <item
         android:id="@+id/passengersFragment"
         app:showAsAction="always"
         android:orderInCategory="2"
-        android:icon="@android:drawable/ic_dialog_map"
+        android:icon="@drawable/ic_avatar"
         android:title="Pasajeros" />
 </menu>
